Revert "Merge branch 'master' of https://github.com/gonaroo/OOA-Gantenbein-Brunner-Patrut-Troller"
This reverts commit 633d367d16709ff25a19f017a1a09b7c098a5c3c, reversing
changes made to 167e775a3a26dfa44335f4e072231bdf4d423b17.

+++ src/user/Kunde.java
@@ -5,9 +5,7 @@ import src.rtd.*;
 
 public class Kunde extends User {
 
-	
-	//muss double sein, weil von Anlage importiert!!
-	public double place(int value) {
+	public int place(int value) {
 		// if address 0 then RTDR
 
 		// or
